+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * **Buy things by their names**
 * **Raid things by their names**
+* Embed IPython
 * Switch back to plain old `requirements.txt`
 * Fix `--help`
 * Use `pydantic` for response classes